+++ src/Network/Network.Test/ScenarioTests/VirtualNetworkGatewayFailoverTests.cs
@@ -1,0 +1,70 @@
+// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using Microsoft.Azure.Commands.Network.Test.ScenarioTests;
+using Microsoft.WindowsAzure.Commands.ScenarioTest;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace Commands.Network.Test.ScenarioTests
+{
+    public class VirtualNetworkGatewayFailoverTests : NetworkTestRunner
+    {
+        public VirtualNetworkGatewayFailoverTests(ITestOutputHelper output)
+            : base(output)
+        {
+        }
+
+        [Fact(Skip = "Requires pre-created virtual network gateway resources")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.exrdev)]
+        public void TestVirtualNetworkGatewayResiliencyInformation()
+        {
+            TestRunner.RunTestScript("Test-VirtualNetworkGatewayResiliencyInformation");
+        }
+
+        [Fact(Skip = "Requires pre-created virtual network gateway resources")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.exrdev)]
+        public void TestVirtualNetworkGatewayRoutesInformation()
+        {
+            TestRunner.RunTestScript("Test-VirtualNetworkGatewayRoutesInformation");
+        }
+
+        [Fact(Skip = "Requires pre-created virtual network gateway resources")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.exrdev)]
+        public void TestStartVirtualNetworkGatewaySiteFailoverTest()
+        {
+            TestRunner.RunTestScript("Test-StartAzureVirtualNetworkGatewaySiteFailoverTest");
+        }
+
+        // Test for StopAzureVirtualNetworkGatewaySiteFailoverTest cmdlet
+        [Fact(Skip = "Requires pre-created virtual network gateway resources")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.exrdev)]
+        public void TestStopVirtualNetworkGatewaySiteFailoverTest()
+        {
+            TestRunner.RunTestScript("Test-StopAzureVirtualNetworkGatewaySiteFailoverTest");
+        }
+
+        [Fact(Skip = "Requires pre-created virtual network gateway resources")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.exrdev)]
+        public void TestCombinedVirtualNetworkGatewayFailoverTestDetails()
+        {
+            TestRunner.RunTestScript("Test-CombinedAzureVirtualNetworkGatewayFailoverTestDetails");
+        }
+	}
+}