+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,12 @@
 --->
 
 ## Upcoming Release
+* Added cmdlet `Get-AzVirtualNetworkGatewayFailoverSingleTestDetails` to retrieve details of a single failover test on a virtual network gateway.
+* Added cmdlet `Get-AzVirtualNetworkGatewayFailoverAllTestsDetails` to list all failover tests for a virtual network gateway.
+* Added cmdlet `Start-AzVirtualNetworkGatewaySiteFailoverTest` to initiate a site failover test on a virtual network gateway.
+* Added cmdlet `Stop-AzVirtualNetworkGatewaySiteFailoverTest` to stop an ongoing site failover test.
+* Added cmdlet `Get-AzVirtualNetworkGatewayResiliencyInformation` to retrieve resiliency information for a virtual network gateway.
+* Added cmdlet `Get-AzVirtualNetworkGatewayRoutesInformation` to get route information for a virtual network gateway.
 * Added new cmdlet for List NetworkSecurityPerimeter ServiceTags
     - `Get-AzNetworkSecurityPerimeterServiceTag`
 * Added properties 'DedicatedBackendConnection', 'ValidateCertChainAndExpiry', 'ValidateSNI', and 'SniName' to Application Gateway Backend HTTP Settings, as well as support for them in the following cmdlets:
@@ -38,7 +44,6 @@
         - `New-AzApplicationGatewayFirewallPolicy`
 * Onboarded FileShares to Private Link Common Cmdlets
 * Updated Api version for StorageSync to Private Link Common Cmdlets
-
 
 ## Version 7.19.0
 * Returned appgw and agc in waf policy